user checking is added

+++ controllers/transactions/transactions.controller.js
@@ -1,10 +1,15 @@
+const createError = require("http-errors");
 const {
   addTransaction,
   deleteTransaction,
 } = require("../../services/transactionsService");
 
-const addTransactionController = async (req, res) => {
+const addTransactionController = async (req, res, next) => {
+  if (!req.user) return next(createError(404, "No users found"));
+  if (!req.user.token) return next(createError(401, "Not authorized"));
+
   const { _id: owner } = req.user;
+
   const { userId, _id, dateTransaction, income, sum, category, description } =
     await addTransaction(req.body, owner);
 
@@ -20,8 +25,12 @@ const addTransactionController = async (req, res) => {
 };
 
 const deleteTransactionController = async (req, res, next) => {
+  if (!req.user) return next(createError(404, "No users found"));
+  if (!req.user.token) return next(createError(401, "Not authorized"));
+
   const { transactionId } = req.params;
   const { _id: owner } = req.user;
+
   const isTransactionDeleted = await deleteTransaction(
     transactionId,
     owner,
